+++ packages/dialtone-vue3/components/toggle/toggle.vue
@@ -1,5 +1,8 @@
 <template>
-  <div class="d-toggle-wrapper">
+  <div
+    class="d-toggle-wrapper"
+    v-bind="addClassStyleAttrs($attrs)"
+  >
     <label
       v-if="hasSlotContent($slots.default)"
       :class="labelClass"
@@ -30,7 +33,7 @@
 
 <script>
 import { warn } from 'vue';
-import { getUniqueString, hasSlotContent } from '@/common/utils';
+import { getUniqueString, hasSlotContent, removeClassStyleAttrs, addClassStyleAttrs } from '@/common/utils';
 import { TOGGLE_CHECKED_VALUES, TOGGLE_SIZE_MODIFIERS } from '@/components/toggle/toggle_constants';
 
 /**
@@ -146,7 +149,7 @@ export default {
   computed: {
     inputListeners () {
       return {
-        ...this.$attrs,
+        ...removeClassStyleAttrs(this.$attrs),
         onClick: _ => this.toggleCheckedValue(),
       };
     },
@@ -183,6 +186,7 @@ export default {
   },
 
   methods: {
+    addClassStyleAttrs,
     toggleCheckedValue () {
       this.$emit('change', !this.internalChecked);