+++ packages/dialtone-vue3/CHANGELOG.md
@@ -1,3 +1,18 @@
+# [3.174.0](https://github.com/dialpad/dialtone/compare/dialtone-vue3/v3.173.2...dialtone-vue3/v3.174.0) (2025-04-07)
+
+
+### Features
+
+* **Editor:** DP-130987 insert inline image through editor ([#692](https://github.com/dialpad/dialtone/issues/692)) ([4616b67](https://github.com/dialpad/dialtone/commit/4616b67e0d8781d932b15200e0f3f7768dd241a7))
+
+## [3.173.2](https://github.com/dialpad/dialtone/compare/dialtone-vue3/v3.173.1...dialtone-vue3/v3.173.2) (2025-04-03)
+
+
+### Bug Fixes
+
+* DLT-2398 synchronize class attrs on vue 3 components ([#688](https://github.com/dialpad/dialtone/issues/688)) ([ff72641](https://github.com/dialpad/dialtone/commit/ff726419f06647ba0d3d6016fddcf698c13ba21d))
+* NO-JIRA $el in vue 3 ([#690](https://github.com/dialpad/dialtone/issues/690)) ([0027ba7](https://github.com/dialpad/dialtone/commit/0027ba7dbe3c394b9185797790a6408a6eae5234))
+
 ## [3.173.1](https://github.com/dialpad/dialtone/compare/dialtone-vue3/v3.173.0...dialtone-vue3/v3.173.1) (2025-04-02)